+++ src/pages/settings/dispatcher.tsx
@@ -1,0 +1,3 @@
+import DispatcherSettings from '@components/Settings/Dispatcher';
+
+export default DispatcherSettings;
